Restore the RCONClient.listen() decorator
